+++ core/genrccom.h
@@ -70,30 +70,30 @@
    Defgeneric                    *LookupDefgenericByMdlOrScope(Environment *,const char *);
    Defgeneric                    *LookupDefgenericInScope(Environment *,const char *);
    Defgeneric                    *GetNextDefgeneric(Environment *,Defgeneric *);
-   unsigned                       GetNextDefmethod(Defgeneric *,unsigned);
+   unsigned short                 GetNextDefmethod(Defgeneric *,unsigned short);
    bool                           DefgenericIsDeletable(Defgeneric *);
-   bool                           DefmethodIsDeletable(Defgeneric *,unsigned);
+   bool                           DefmethodIsDeletable(Defgeneric *,unsigned short);
    void                           UndefgenericCommand(Environment *,UDFContext *,UDFValue *);
    void                           GetDefgenericModuleCommand(Environment *,UDFContext *,UDFValue *);
    void                           UndefmethodCommand(Environment *,UDFContext *,UDFValue *);
    Defmethod                     *GetDefmethodPointer(Defgeneric *,long);
    bool                           Undefgeneric(Defgeneric *,Environment *);
-   bool                           Undefmethod(Defgeneric *,unsigned,Environment *);
+   bool                           Undefmethod(Defgeneric *,unsigned short,Environment *);
 #if ! OBJECT_SYSTEM
    void                           TypeCommand(Environment *,UDFContext *,UDFValue *);
 #endif
 #if DEBUGGING_FUNCTIONS || PROFILING_FUNCTIONS
-   void                           DefmethodDescription(Defgeneric *,unsigned,StringBuilder *);
+   void                           DefmethodDescription(Defgeneric *,unsigned short,StringBuilder *);
 #endif
 #if DEBUGGING_FUNCTIONS
    bool                           DefgenericGetWatch(Defgeneric *);
    void                           DefgenericSetWatch(Defgeneric *,bool);
-   bool                           DefmethodGetWatch(Defgeneric *,unsigned);
-   void                           DefmethodSetWatch(Defgeneric *,unsigned,bool);
+   bool                           DefmethodGetWatch(Defgeneric *,unsigned short);
+   void                           DefmethodSetWatch(Defgeneric *,unsigned short,bool);
    void                           PPDefgenericCommand(Environment *,UDFContext *,UDFValue *);
    void                           PPDefmethodCommand(Environment *,UDFContext *,UDFValue *);
    void                           ListDefmethodsCommand(Environment *,UDFContext *,UDFValue *);
-   const char                    *DefmethodPPForm(Defgeneric *,unsigned);
+   const char                    *DefmethodPPForm(Defgeneric *,unsigned short);
    void                           ListDefgenericsCommand(Environment *,UDFContext *,UDFValue *);
    void                           ListDefgenerics(Environment *,const char *,Defmodule *);
    void                           ListDefmethods(Environment *,const char *,Defgeneric *);
@@ -103,7 +103,7 @@
    void                           GetDefmethodListCommand(Environment *,UDFContext *,UDFValue *);
    void                           GetDefmethodList(Environment *,CLIPSValue *,Defgeneric *);
    void                           GetMethodRestrictionsCommand(Environment *,UDFContext *,UDFValue *);
-   void                           GetMethodRestrictions(Defgeneric *,unsigned,CLIPSValue *);
+   void                           GetMethodRestrictions(Defgeneric *,unsigned short,CLIPSValue *);
    void                           SetNextDefgeneric(Defgeneric *,Defgeneric *);
    const char                    *DefgenericModule(Defgeneric *);
    const char                    *DefgenericName(Defgeneric *);
