+++ core/memalloc.h
@@ -175,8 +175,8 @@ struct memoryPtr
 
 struct memoryData
   {
-   long MemoryAmount;
-   long MemoryCalls;
+   long long MemoryAmount;
+   long long MemoryCalls;
    bool ConserveMemory;
    OutOfMemoryFunction *OutOfMemoryCallback;
    struct memoryPtr *TempMemoryPtr;
@@ -192,16 +192,16 @@ struct memoryData
    OutOfMemoryFunction           *SetOutOfMemoryFunction(Environment *,OutOfMemoryFunction *);
    void                           genfree(Environment *,void *,size_t);
    void                          *genrealloc(Environment *,void *,size_t,size_t);
-   long                           MemUsed(Environment *);
-   long                           MemRequests(Environment *);
-   long                           UpdateMemoryUsed(Environment *,long);
-   long                           UpdateMemoryRequests(Environment *,long);
+   long long                      MemUsed(Environment *);
+   long long                      MemRequests(Environment *);
+   long long                      UpdateMemoryUsed(Environment *,long long);
+   long long                      UpdateMemoryRequests(Environment *,long long);
    long                           ReleaseMem(Environment *,long);
    void                          *gm1(Environment *,size_t);
    void                          *gm2(Environment *,size_t);
    void                           rm(Environment *,void *,size_t);
-   unsigned long                  PoolSize(Environment *);
-   unsigned long                  ActualPoolSize(Environment *);
+   size_t                         PoolSize(Environment *);
+   size_t                         ActualPoolSize(Environment *);
    void                          *RequestChunk(Environment *,size_t);
    int                            ReturnChunk(Environment *,void *,size_t);
    bool                           SetConserveMemory(Environment *,bool);
